Refactoring old classes to new framework
- CoreApp extracted from index.php and moved to \IcarosNet\LastHammer\CoreApp
- ClassManager moved to \IcarosNet\LastHammer\Manager
- autoloder moved to apps/core/ folder
- index.php cleaned
