+++ apps/Classes/Gen/Date.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace IcarosNet\LastHammer\Gen;
-use CoreApp;
+use IcarosNet\LastHammer\CoreApp;
 use Exception;
 
 class Date
