+++ apps/Classes/Gen/Device.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace IcarosNet\LastHammer\Gen;
-use CoreApp;
+use IcarosNet\LastHammer\CoreApp;
 use Mobile_Detect;
 
 class Device
