+++ apps/Classes/Gen/ExecTime.php
@@ -1,7 +1,9 @@
 <?php
 
 namespace IcarosNet\LastHammer\Gen;
-use CoreApp;
+
+
+use IcarosNet\LastHammer\CoreApp;
 
 class ExecTime
 {
