+++ apps/Classes/Gen/Log.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace IcarosNet\LastHammer\Gen;
-use CoreApp;
+use IcarosNet\LastHammer\CoreApp;
 
 class Log
 {
