+++ apps/Classes/Gen/Session.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace IcarosNet\LastHammer\Gen;
-use CoreApp;
+use IcarosNet\LastHammer\CoreApp;
 
 class Session
 {
