+++ apps/Classes/Mvc/OutputData.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace IcarosNet\LastHammer\Mvc;
-use CoreApp;
+use IcarosNet\LastHammer\CoreApp;
 use IcarosNet\LastHammer\Gen\Vars;
 
 class OutputData
