+++ apps/Classes/Mvc/Process.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace IcarosNet\LastHammer\Mvc;
-use CoreApp;
+use IcarosNet\LastHammer\CoreApp;
 use Exception;
 
 class Process
